Move the preview code to its own module and add hooks
